fix(desktop): Phase 12 round 4 — Long.dp, Long hex colors, MenuItem.label, noinline
Run 33943041377: step 5 (Android: crash fix + FGS service) is GREEN;
step 6 finally compile-checked the desktop code — 8 small errors:
- Main.kt: Long has no .dp (Int/Float/Double do) — persist geometry as
  px Longs, convert via toFloat() before .dp for rememberWindowState
- TrayIcons.kt: 0xFF161616/0xFFBB86FC are Long literals — .toInt() for
  Color(Int); Float.coerceAtLeast(1) — .toInt() first
- DhunTray.kt: java.awt.MenuItem uses label, not text; inline lambda
  escaping to SwingUtilities.invokeLater needs noinline
Details + the settled 1.8.2 API facts (application{} IS composable;
Window(undecorated=…) overload has all defaults) in .ai/DEBUG_LOG.md.

+++ app-desktop/src/jvmMain/kotlin/dev/dhun/desktop/native/TrayIcons.kt
@@ -15,9 +15,9 @@ import java.awt.image.BufferedImage
  */
 object TrayIcons {
 
-    private val background = Color(0xFF161616)
+    private val background = Color(0xFF161616.toInt())
     private val border = Color(0x33FFFFFF)
-    private val accent = Color(0xFFBB86FC)
+    private val accent = Color(0xFFBB86FC.toInt())
 
     fun playing(size: Int = 32): BufferedImage = draw(size) { g, s ->
         tile(g, s)
@@ -36,8 +36,8 @@ object TrayIcons {
     fun paused(size: Int = 32): BufferedImage = draw(size) { g, s ->
         tile(g, s)
         g.color = accent
-        val barW = (s * 0.12f).coerceAtLeast(1)
-        val gap = (s * 0.09f).coerceAtLeast(1)
+        val barW = (s * 0.12f).toInt().coerceAtLeast(1)
+        val gap = (s * 0.09f).toInt().coerceAtLeast(1)
         val h = (s * 0.40f).toInt()
         val top = (s - h) / 2
         val mid = s / 2
